backend/requirements.txt
# requirements.txt şöyle görünmeli:
Django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
psycopg2-binary
python-dotenv
google-auth
google-auth-oauthlib
google-api-python-client
django-extensions